+++ tcob/src/audio/SoundGenerator.cpp
@@ -393,7 +393,7 @@ auto sound_generator::create_buffer(sound_wave const& wave) -> buffer
     // NOTE: We reserve enough space for up to 10 seconds of wave audio at given sample rate
     // By default we use f32 size samples, they are converted to desired sample size at the end
     std::vector<f32> samples;
-    samples.resize(MAX_WAVE_LENGTH_SECONDS * wave.SampleRate);
+    samples.resize(static_cast<usize>(MAX_WAVE_LENGTH_SECONDS * wave.SampleRate));
 
     bool generatingSample {true};
     i32  sampleCount {0};
@@ -484,10 +484,10 @@ auto sound_generator::create_buffer(sound_wave const& wave) -> buffer
         ssample = (ssample / MAX_SUPERSAMPLING) * SAMPLE_SCALE_COEFICIENT;
 
         // Accumulate samples in the buffer
-        samples[i] = std::clamp(ssample, -1.0f, 1.0f);
+        samples[static_cast<usize>(i)] = std::clamp(ssample, -1.0f, 1.0f);
     }
 
-    samples.resize(sampleCount);
+    samples.resize(static_cast<usize>(sampleCount));
     return buffer::Create({.Channels = 1, .SampleRate = wave.SampleRate, .FrameCount = std::ssize(samples)}, samples);
 }
