+++ tcob/src/gfx/backend/gles30/GLESTexture.cpp
@@ -130,7 +130,7 @@ auto gl_texture::copy_to_image(u32 depth) const -> image
     bind();
 
     std::vector<ubyte> buffer;
-    buffer.resize(_size.Width * _size.Height * 4);
+    buffer.resize(static_cast<usize>(_size.Width * _size.Height * 4));
 
     gl_framebuffer frameBuffer;
     frameBuffer.attach_texture(this, depth);
